fix style in requests_test, and move acceptance test to a ruletypes package
