+++ acceptance/openstack/networking/v2/extensions/qos/ruletypes/ruletypes_test.go
@@ -1,4 +1,4 @@
-package qos
+package ruletypes
 
 import (
 	"testing"
