+++ openstack/networking/v2/extensions/qos/ruletypes/testing/requests_test.go
@@ -36,6 +36,6 @@ func TestListRuleTypes(t *testing.T) {
 		return
 	}
 
-	expected := []ruletypes.RuleType{{"bandwidth_limit"}, {"dscp_marking"}, {"minimum_bandwidth"}}
+	expected := []ruletypes.RuleType{{Type: "bandwidth_limit"}, {Type: "dscp_marking"}, {Type:"minimum_bandwidth"}}
 	th.AssertDeepEquals(t, expected, rules)
 }
